Allow custom timeout value for webhook calls
Because webhooks add to API request latency, they should evaluate as
quickly as possible. timeoutSeconds allows configuring how long the
API server should wait for a webhook to respond before treating the call
as a failure.

The default values from Kubernetes are, however, too large. For
admissionregistration.k8s.io/v1 the default value is 10 seconds while
for admissionregistration.k8s.io/v1beta1 is 30 seconds.

We keep 30 seconds as this was the original unconfigurable value.

Signed-off-by: Renan Gonçalves <renan.saddam@gmail.com>